assets_py_unpacker/requirements.txt
Cython>=3.0.10
ffmpeg_python>=0.2.0
PyCriCodecs>=0.4.8
numpy>=1.23.5
opencv-python>=4.8.1.78
